pyproject.toml
[project]
name = "py-sadl"
version = "1.3.0"
description = "Simple Autograd Deep Learning: A minimal, readable deep learning framework built on NumPy"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Tim Cares" }]
keywords = [
    "deep-learning",
    "machine-learning",
    "autograd",
    "automatic-differentiation",
    "neural-networks",
    "numpy",
    "cupy",
    "educational",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: GPU :: NVIDIA CUDA",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Mathematics",
    "Typing :: Typed",
]

dependencies = [
    "numpy>=2.4.0",
]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "python-semantic-release>=9.0.0",
    "commitizen>=4.13.9",
    "pre-commit>=4.0.0",
    "interrogate>=1.7.0",
    "pip-audit>=2.10.0",
    "pyright>=1.1.408",
]

gpu = [
    "cupy-cuda12x>=13.0.0",
]

notebook = [
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.8",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "pillow>=12.1.1", # for CVE fix for transitive matplotlib dependency
    "notebook>=7.5.3",
    "datasets>=4.5.0",
]

build = ["uv ~= 0.10.0"]

[project.urls]
Homepage = "https://github.com/timcares/sadl"
Documentation = "https://github.com/timcares/sadl#readme"
Repository = "https://github.com/timcares/sadl"
Issues = "https://github.com/timcares/sadl/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["sadl"]

# =============================================================================
# Ruff Configuration
# =============================================================================
[tool.ruff]
target-version = "py311"
line-length = 100
src = ["sadl"]
exclude = ["notebooks/", "tests/"]

[tool.ruff.lint]
select = [
    # --- Core ---
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    # --- Bug / Code-smell ---
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
    "PL",     # pylint
    "RUF",    # ruff-specific
    "PIE",    # flake8-pie
    "PERF",   # perflint
    "RET",    # flake8-return
    # --- Documentation / Typing ---
    "D",      # pydocstyle
    "ANN",    # flake8-annotations
    "TC",     # flake8-type-checking
    # --- Security ---
    "S",      # flake8-bandit (replaces standalone bandit)
    # --- Style ---
    "ARG",    # flake8-unused-arguments
    "A",      # flake8-builtins (prevent shadowing builtins)
    "ISC",    # flake8-implicit-str-concat
    "T20",    # flake8-print (catch stray print() calls)
    "PTH",    # flake8-use-pathlib
    "DTZ",    # flake8-datetimez
    "RSE",    # flake8-raise
    "TRY",    # tryceratops (exception handling)
    "LOG",    # flake8-logging
    # --- Testing ---
    "PT",     # flake8-pytest-style
]
ignore = [
    "E501",   # line length — handled by ruff formatter
    "D100",   # module docstrings — interrogate covers doc coverage
    "D104",   # __init__.py docstrings — unnecessary boilerplate
    "TRY003", # long exception messages — too strict for real-world code
    "ISC001", # implicit string concat single-line — conflicts with ruff formatter
    "PTH123", # Unnceessary verbose "Path.open" instead of "open" for files
    "ANN401", # Any not allowed for *args and **kwargs, leads to awkward code
    "A001",   # Shadowing python buildin attributes is ok, since we use a lot of familiar functions like "max", "sum", ...
    "A004",   # see "A001"
    "D105",   # No docstrings needed in magic methods like __sub__, as they are usually self-explanatory
]

[tool.ruff.lint.pylint]
max-args = 10
max-returns = 8

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true


# =============================================================================
# Interrogate — Docstring Coverage
# =============================================================================

[tool.interrogate]
paths = ["sadl"]
fail-under = 90
ignore-init-method = true
ignore-init-module = true
ignore-module = true
exclude = ["tests"]


# =============================================================================
# Pyright — Static Type Checking
# =============================================================================

[tool.pyright]
typeCheckingMode = "strict"
venvPath = "."
venv = ".venv"
include = ["sadl"]

reportUnnecessaryIsInstance = false # needed for untyped client code passing in invalid objects to the api

# pyright incorrectly handles "BACKEND" var
reportConstantRedefinition = false
reportUnnecessaryComparison = false

# =============================================================================
# Pytest Configuration
# =============================================================================
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "-ra",
    "-v",
    "--tb=short",
    "--durations=10",
]


# =============================================================================
# Coverage
# =============================================================================

[tool.coverage.run]
source = ["sadl"]
branch = true
omit = ["tests/*", "*/__init__.py"]

[tool.coverage.report]
fail_under = 80
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "@abstractmethod",
    'if __name__ == "__main__"',
]

# =============================================================================
# Semantic Release Configuration
# =============================================================================
[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
# The build command for semantic release is more than just the actual
#   build command "uv build", because the package version in
#   the "uv.lock" file will become stale.
#   See: https://python-semantic-release.readthedocs.io/en/latest/configuration/configuration-guides/uv_integration.html#updating-the-uv-lock
build_command = """
    python -m pip install -e '.[build]'
    uv lock --upgrade-package "$PACKAGE_NAME"
    git add uv.lock
    uv build
"""
commit_parser = "conventional"
commit_message = "chore(release): v{version}"
tag_format = "v{version}"

[tool.semantic_release.changelog.default_templates]
changelog_file = "CHANGELOG.md"
output_format = "md"

[tool.semantic_release.branches.main]
match = "(main|master)"
prerelease = false

[tool.semantic_release.branches.alpha]
match = ".*"
prerelease_token = "alpha"
prerelease = true

[tool.semantic_release.remote]
type = "github"
name = "origin"

[tool.semantic_release.remote.token]
env = "GH_TOKEN"


# =============================================================================
# Commitizen — Local commit message format enforcement only
# =============================================================================

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "v$version"
